requirements.txt

flask==3.0.3
requests==2.32.2
python-dotenv==1.0.1